+++ backend/src/configs/router.ts
@@ -1,9 +1,12 @@
 import { Router } from "express";
-import { authRouter } from "../auth/auth.route.js";
-import { searchRouter } from "../search/search.route.js";
-import { mediaRouter } from "../media/media.route.js";
-import { adminRouter } from "../admin/admin.route.js";
-import { userRouter } from "../user/user.route.js";
+import { authRouter } from "../api/auth/auth.route.js";
+import { searchRouter } from "../api/search/search.route.js";
+import { mediaRouter } from "../api/media/media.route.js";
+import { adminRouter } from "../api/admin/admin.route.js";
+import { userRouter } from "../api/user/user.route.js";
+import { reviewsRouter } from "../api/reviews/reviews.route.js";
+import { watchRouter } from "../api/watch/watch.route.js";
+import { genreRouter } from "../api/genre/genre.route.js";
 
 const router = Router();
