+++ ui/src/pages/settingsPage/settingsPage.tsx
@@ -1,0 +1,143 @@
+import { Search } from "lucide-react";
+import { useState } from "react";
+import { Button } from "@/components/ui/button";
+import { getUserSettings, setUserSettings } from "@/api/user";
+import { useMutation, useQuery } from "@tanstack/react-query";
+import { ProfileSettingsCategoryProfile } from "./components/CategoryProfile";
+import { ProfileSettingsCategoryAccount } from "./components/CategoryAccount";
+import { ProfileSettingsCategoryAppearance } from "./components/CategoryAppearance";
+import { ProfileSettingsCategoryActivity } from "./components/CategoryActivity";
+import { ProfileSettingsCategoryHelpAndSupport } from "./components/CategoryHelpAndSupport";
+
+export function ProfileSettings(){
+    const [changedSettings, setChangedSettings] = useState<Map<string,string>>(new Map<string,string>());
+
+    const [selectedMenu, setSelectedMenu] = useState("Profile");
+
+    const { data:currentSettings, refetch:refetchCurrentSettings } = useQuery({
+        queryKey: ['profileSettingsCurrentSettings'],
+        queryFn: async () => {
+            const a = await getUserSettings();
+            const b = new Map<string,string>();
+            for (const [key, value] of Object.entries(a)) {
+                b.set(key,String(value));
+            }
+            const c = new Map<string,string>();
+            changedSettings.forEach((value:string,key:string) => {
+                if (value != b.get(key)){
+                    c.set(key,value);
+                }
+            })
+            setChangedSettings(c);
+            return b;
+        }
+    });
+
+    const { mutate:applyChangedSettings } = useMutation({
+        mutationFn: async () => {
+            await setUserSettings(changedSettings)
+            refetchCurrentSettings();
+            setChangedSettings(new Map<string,string>());
+        }
+    });
+
+    const updateSetting = (key:string, value:string) => {
+        const currentValue = currentSettings?.get(key);
+        const map:Map<string,string> = new Map<string,string>();
+        changedSettings.forEach((old_value:string,old_key:string)=>{
+            if (old_key == key) {
+                if (value != currentValue){
+                    map.set(old_key,value);
+                }
+            }
+            else {
+                map.set(old_key,old_value);
+            }
+        })
+        if (!changedSettings.has(key) && (value != currentValue)) {
+            map.set(key,value);
+        }
+        setChangedSettings(map);
+    }
+
+    const findSetting = (key:string) => {
+        const changedValue = changedSettings.get(key);
+        if (changedValue != null) {
+            return changedValue;
+        }
+        const currentValue = currentSettings?.get(key);
+        if (currentValue != null) {
+            return currentValue;
+        }
+        return null;
+    }
+
+    console.log(currentSettings, changedSettings);
+
+    return (
+    <>
+        <div className="flex items-start rounded-3xl w-full min-h-[calc(100vh-100px)] bg-black border-white border">
+            <ProfileSettingsMenu selectedMenu={selectedMenu} setSelectedMenu={setSelectedMenu}></ProfileSettingsMenu>
+            <div className=" bg-white w-[1px] self-stretch"></div>
+            <div className="flex p-5 flex-col items-start gap-3">
+                <div className="flex flex-col items-start gap-1 self-stretch">
+                    <p className="text-white text-2xl font-light leading-normal">
+                        {selectedMenu}
+                    </p>
+                    <div className="self-stretch h-px bg-[#7F7F7E]"></div>
+                    {selectedMenu=="Profile"?<ProfileSettingsCategoryProfile updateSetting={updateSetting} findSetting={findSetting}></ProfileSettingsCategoryProfile>:null}
+                    {selectedMenu=="Account"?<ProfileSettingsCategoryAccount updateSetting={updateSetting} findSetting={findSetting}></ProfileSettingsCategoryAccount>:null}
+                    {selectedMenu=="Appearance"?<ProfileSettingsCategoryAppearance updateSetting={updateSetting} findSetting={findSetting}></ProfileSettingsCategoryAppearance>:null}
+                    {selectedMenu=="Activity"?<ProfileSettingsCategoryActivity updateSetting={updateSetting} findSetting={findSetting}></ProfileSettingsCategoryActivity>:null}
+                    {selectedMenu=="Help & Support"?<ProfileSettingsCategoryHelpAndSupport updateSetting={updateSetting} findSetting={findSetting}></ProfileSettingsCategoryHelpAndSupport>:null}
+                </div>
+            </div>
+        </div>
+        {
+        changedSettings.size > 0
+        ?(
+            <div className="flex items-center justify-center fixed bottom-5 p-3 gap-3 bg-black border border-white rounded-2xl">
+                <p className="flex self-center text-base font-normal">
+                    {changedSettings.size} {changedSettings.size != 1?"settings have been modified":"setting has been modified"}
+                </p>
+                <Button onClick={()=>{applyChangedSettings()}} variant={"default"}>Apply Changes</Button>
+            </div>
+        ):null}
+    </>
+    )
+}
+
+function ProfileSettingsMenu({selectedMenu, setSelectedMenu}:{selectedMenu:string,setSelectedMenu:(a:string)=>void}){
+    return (
+        <div className="flex flex-col min-w-[33%] flex-shrink-0">
+            <div className="flex py-5 px-3 justify-between items-center self-stretch border-b border-white">
+                <p className="text-white text-2xl font-light leading-normal">
+                    Settings
+                </p>
+                <Search className="w-[21px] h-[21px]"></Search>
+            </div>
+            <div className="flex flex-col items-start self-stretch">
+                <ProfileSettingsMenuButton selectedMenu={selectedMenu} setSelectedMenu={setSelectedMenu} category="Profile"></ProfileSettingsMenuButton>
+                <line className="self-stretch h-[1px] bg-[#7F7F7E]"></line>
+                <ProfileSettingsMenuButton selectedMenu={selectedMenu} setSelectedMenu={setSelectedMenu} category="Account"></ProfileSettingsMenuButton>
+                <line className="self-stretch h-[1px] bg-[#7F7F7E]"></line>
+                <ProfileSettingsMenuButton selectedMenu={selectedMenu} setSelectedMenu={setSelectedMenu} category="Appearance"></ProfileSettingsMenuButton>
+                <line className="self-stretch h-[1px] bg-[#7F7F7E]"></line>
+                <ProfileSettingsMenuButton selectedMenu={selectedMenu} setSelectedMenu={setSelectedMenu} category="Activity"></ProfileSettingsMenuButton>
+                <line className="self-stretch h-[1px] bg-[#7F7F7E]"></line>
+                <ProfileSettingsMenuButton selectedMenu={selectedMenu} setSelectedMenu={setSelectedMenu} category="Help & Support"></ProfileSettingsMenuButton>
+                <line className="self-stretch h-[1px] bg-[#7F7F7E]"></line>
+            </div>
+        </div>
+    )
+}
+
+function ProfileSettingsMenuButton({category, selectedMenu, setSelectedMenu}: {category:string,selectedMenu:string,setSelectedMenu:(a:string)=>void}){
+    return (
+        <div onClick={()=>{setSelectedMenu(category)}} className={"flex py-5 px-3 items-center gap-3 self-stretch border-r-8 border-solid hover:bg-neutral-900 " + (category == selectedMenu ? " border-purple" : " border-[#7F7F7E]")}>
+            <p className="text-white text-base font-medium leading-normal">
+                {category}
+            </p>
+        </div>
+    )
+}
